requirements.txt
pandas
pytest
streamlit
scikit-learn
google-genai
numpy
python-dotenv
langchain
langchain-google-genai
